+++ makedb.py
@@ -2,6 +2,8 @@ import os
 from model.Board import Board
 from model.Slip import gen_slip
 from model.Tag import Tag
+import model.Media
+import model.Poster
 from shared import db
 
 db.create_all()
