+++ model/PostReplyExtension.py
@@ -4,8 +4,6 @@ from flask import url_for
 from markdown import Extension
 
 from model.PostReplyPattern import PostReplyPattern
-from model.Thread import Thread
-
 
 
 class PostReplyExtension(Extension):
