+++ model/PostReplyPattern.py
@@ -1,9 +1,17 @@
 from xml import etree
 
+from flask import url_for
+from markdown.util import etree
 from markdown.inlinepatterns import Pattern
 
-from model.Post import url_for_post
+#from .Post import url_for_post
 from model.Reply import REPLY_REGEXP
+
+
+def url_for_post(post_id):
+    from model.Thread import Thread
+    thread = Thread.query.filter(Thread.posts.any(id=post_id)).one()
+    return url_for("view_thread", thread_id=thread.id) + "#" + str(post_id)
 
 
 class PostReplyPattern(Pattern):
