+++ model/Slip.py
@@ -1,6 +1,7 @@
 import uuid
 
 from flask import session
+from werkzeug.security import generate_password_hash, check_password_hash
 
 from model.Session import Session
 from shared import db
