+++ templates/catalog-view.html
@@ -58,13 +58,13 @@
         {% endif %}
         <div class="thread border-top rounded">
             {% if thread["media"] %}
-            <a href="{{ url_for(" view_thread", thread_id=thread["id"]) }}">
-            <img class="border-left border-right w-100 rounded-top" src="{{ url_for(" uploaded_thumb",
+            <a href="{{ url_for("view_thread", thread_id=thread["id"]) }}">
+            <img class="border-left border-right w-100 rounded-top" src="{{ url_for("uploaded_thumb",
             media_id=thread["media"]) }}">
             </a>
             {% endif %}
             {% if thread["subject"] %}
-            <div class="{{ " p-2 border-left border-right
+            <div class="{{ "p-2 border-left border-right
             " + ns.bg_style }}">
             <h5 class="{{ ns.text_style }}">{{ thread["subject"] }}</h5>
         </div>
@@ -76,10 +76,10 @@
         <div class="pb-2 pt-2 pl-4 pr-4 border-left border-right">
             {{ thread["body"]|safe|truncate() }}
         </div>
-        <div class="{{ " pt-1 border border-top-0 rounded-bottom
+        <div class="{{ "pt-1 border border-top-0 rounded-bottom
         " + ns.bg_style }}">
-        <a href="{{ url_for(" view_thread", thread_id=thread["id"]) }}" class="btn btn-dark m-2">View thread</a>
-        <span class="{{ " mt-3 mr-2 float-right " + ns.text_style }}">
+        <a href="{{ url_for("view_thread", thread_id=thread["id"]) }}" class="btn btn-dark m-2">View thread</a>
+        <span class="{{ "mt-3 mr-2 float-right " + ns.text_style }}">
         <i class="fas fa-comment"></i> {{ thread["num_replies"] }}
         <i class="fas fa-image"></i> {{ thread["num_media"] }}
         </span>
