+++ templates/faq.html
@@ -42,7 +42,7 @@ frequently asked questions
         can do, in addition to gracefully downgrading if Javascript has been disabled.</p>
     <h1 id="what-topics">What topics can I talk about here?</h1>
     <p class="lead">Plenty!</p>
-    <p>Go look at the <a href="{{ url_for(" list_boards") }}">board listing</a> to get an idea of the sort of
+    <p>Go look at the <a href="{{ url_for("list_boards") }}">board listing</a> to get an idea of the sort of
         things you can talk about.</p>
 </div>
 {% endblock %}
