+++ templates/gallery.html
@@ -36,8 +36,8 @@ media gallery
 <script src="https://unpkg.com/masonry-layout@4/dist/masonry.pkgd.min.js"></script>
 {% endblock %}
 {% block navbar %}
-<a href="{{ url_for(" view_thread", thread_id=thread_id) }}" class="nav-item nav-link">Back to thread</a>
-<a href="{{ url_for(" view_catalog", board_id=board_id) }}" class="nav-item nav-link">Catalog</a>
+<a href="{{ url_for("view_thread", thread_id=thread_id) }}" class="nav-item nav-link">Back to thread</a>
+<a href="{{ url_for("view_catalog", board_id=board_id) }}" class="nav-item nav-link">Catalog</a>
 {% endblock %}
 {% block content %}
 <div class="container-fluid">
@@ -51,8 +51,8 @@ media gallery
         {% for post in posts %}
         {% if post["media"] %}
         <div class="gallery-entry">
-            <a href="{{ url_for(" uploaded_file", media_id=post["media"]) }}">
-            <img class="img-fluid img-thumbnail" src="{{ url_for(" uploaded_thumb", media_id=post["media"]) }}">
+            <a href="{{ url_for("uploaded_file", media_id=post["media"]) }}">
+            <img class="img-fluid img-thumbnail" src="{{ url_for("uploaded_thumb", media_id=post["media"]) }}">
             </a>
             {% if post["media_ext"] in ["gif", "webm"] %}
             <small class="text-muted">Animated - click to play</small>
