+++ templates/index.html
@@ -8,9 +8,9 @@
             <p class="lead">An anonymous imageboard for the 21st century</p>
             <hr class="my-4">
             <p>If you're not sure what an imageboard is, or why you'd want to use one,
-                check out the <a href="{{ url_for(" faq") + "#whats-an-imageboard" }}">FAQ</a>.
+                check out the <a href="{{ url_for("faq") + "#whats-an-imageboard" }}">FAQ</a>.
                 If you already know what an imageboard is, but aren't quite aware of what
-                Maniwani brings to the table, there's another <a href="{{ url_for(" faq") + "#why-maniwani" }}">FAQ
+                Maniwani brings to the table, there's another <a href="{{ url_for("faq") + "#why-maniwani" }}">FAQ
                 entry</a>
                 that should help you out. If you want to dive right in, you can see the most
                 recently-updated threads on the site below.</p>
